src/requirements.txt
langgraph
python-dotenv
pandas
langchain_groq
groq
tiktoken
crewai
pyautogen